pulled out some input and platform code
